hello ae

+++ IEmployee.cs
@@ -1,5 +1,5 @@
 using System;
 interface IEmployee
 {
-    public int CalcSalary();
+    public double CalcSalary();
 }
